Harden int truncation in shape-derived counters
Closes XOP21.  Twenty-four sites that cast TShape::Size() or workspace
sizes to `int` without a guard are now either promoted to int64_t /
mxnet::index_t / size_t, or narrowed only after an explicit CHECK_LE:

- LayerNorm / GroupNorm channel_size promoted to int64_t in the header,
  CPU MKL path, and CUDA backward decl so the wider type survives the
  template/Implmentation hand-off.
- PSROIPool, ROIAlign forward/backward count and the internal CPU helper
  signatures all switched to mxnet::index_t, with the OpenMP loop
  variable promoted too.
- depthwise_convolution forward + data-grad + weight-grad now hold the
  shape product as index_t, CHECK_LE(INT_MAX) before narrowing to int at
  the CUDA kernel launch, and the block_num computation is done in
  index_t before its own narrowing.
- fully_connected bias-add N / M, transformer-inl DiagMM / per-thread
  counts, and allclose num_items use index_t.
- cudnn_bilinear_sampler and cudnn_spatial_transformer add CHECK_LE per
  dim before writing into the cuDNN descriptor's int[] dim array.
- image_random adds CHECK_LE on the channel reads before narrowing for
  Normalize / ToTensor in both 3-D and 4-D paths.
- linalg_impl: LAPACK workspace size guarded with CHECK_LE, and
  matrix_size / pivot_size / offset bookkeeping promoted to size_t.

Tests: tests/python/unittest/test_xop21_large_shape_validation.py
(4 passed, 1 skipped placeholder for the >INT_MAX no-allocation case).

+++ src/operator/nn/group_norm-inl.h
@@ -123,7 +123,7 @@ void GroupNormCompute(const nnvm::NodeAttrs& attrs,
 
   mxnet::TShape red_src_shape, red_dst_shape;
   BroadcastReduceShapeCompact(temp_data_shape, moments_shape, &red_src_shape, &red_dst_shape);
-  int channel_size = red_src_shape.Size() / red_dst_shape.Size();
+  int64_t channel_size = red_src_shape.Size() / red_dst_shape.Size();
 
   TBlob data_        = data.reshape(red_src_shape);
   const TBlob& mean_ = mean.reshape(red_dst_shape);
